requirements_personal.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-ckeditor==6.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
Faker==5.1.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
